requirements.txt
# Campus Security System - Backend Dependencies
# FastAPI + SQLite with SQLModel

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==12.0

# Database
sqlmodel==0.0.22
sqlalchemy==2.0.35

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# Development
pytest==8.3.3
httpx==0.27.2
